sdk/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "poppie-sdk"
version = "0.1.0"
description = "Python SDK for the poppie TOTP manager"
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.60",
    "protobuf>=4.25",
]

[tool.hatch.build.targets.wheel]
packages = ["src/poppie"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
